refactoring InceptionStructure
