+++ script/model/sklearn_like_model/net_structure/InceptionSructure/Inception_ResnetV1Structure.py
@@ -1,0 +1,8 @@
+from script.model.sklearn_like_model.net_structure.InceptionSructure.BaseInceptionStructure import \
+    BaseInceptionStructure
+
+
+class Inception_ResnetV1Structure(BaseInceptionStructure):
+    def build(self):
+        # TODO
+        raise NotImplementedError
